LociMap Contig: add convenience ctors
